+++ SeriesGuide/src/com/battlelancer/seriesguide/service/OnAlarmReceiver.java
@@ -1,0 +1,17 @@
+
+package com.battlelancer.seriesguide.service;
+
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+
+public class OnAlarmReceiver extends BroadcastReceiver {
+
+    @Override
+    public void onReceive(Context context, Intent intent) {
+        // run the notification service
+        Intent i = new Intent(context, NotificationService.class);
+        context.startService(i);
+    }
+
+}